refactor(langgraph-starter): rename catalog files to match naming convention
schema.ts → demonstrationCatalogDefinitions.ts
catalog.tsx → demonstrationCatalogRenderers.tsx

File names now match the exported symbols.
